Adds CRU routes for physicians. Creates initial paths for public and search pages.
