+++ controllers/public.js
@@ -1,0 +1,12 @@
+var express = require('express');
+var router = express.Router();
+
+router.get('/', function(req, res) {
+
+});
+
+router.get('/login', function(req, res) {
+
+});
+
+module.exports = router;
